django182_requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
certifi==2017.11.5
chardet==3.0.4
Django==1.8.2
django-celery==3.1.17
django-haystack==2.6.0
django-redis==4.8.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.6
itsdangerous==0.24
jieba==0.38
kombu==3.0.37
mutagen==1.39
packaging==16.8
Pillow==3.4.1
PyMySQL==0.7.11
pyparsing==2.2.0
pytz==2016.10
redis==2.10.5
requests==2.18.4
six==1.10.0
urllib3==1.22
uWSGI==2.0.14
Whoosh==2.7.4
